Moved to shared validations
