+++ data/validations/sharedValidations.js
@@ -1,3 +1,15 @@
+const { ObjectId } = require('mongodb');
+
+function stringifyId(id) {
+    return id.toString();
+}
+
+function validateObjectId(id) {
+    if (!ObjectId.isValid(id)) {
+        throw 'invalid object ID';
+    }
+}
+
 function checkArgumentLength(args, length) {
     let argLength = args.length;
 
@@ -8,18 +20,61 @@ function checkArgumentLength(args, length) {
     }
 }
 
+function checkParamPresent(variable, variableName) {
+    if (!variable || variable === undefined) {
+        throw `${variableName || 'provided variable'} is not present`;
+    }
+}
+
 function checkIsString(string, variableName) {
     if (typeof string !== 'string') {
         throw `${variableName || 'provided variable'} is not a string`;
     }
 }
 
-function lower(attr) {
-    return attr.toLowerCase();
+function cleanUpString(string) {
+    return string.trim();
+}
+
+function checkStringLength(string, variableName, minLength) {
+    if (string.length === 0) {
+        throw `${variableName || 'provided variable'} is empty or only contains spaces`;
+    }
+
+    minLength = minLength || 1;
+    if (string.length < minLength) {
+        throw `${variableName || 'provided variable'} must be more than ${minLength} characters`;
+    }
+}
+
+function isValidEmail(email) {
+  // Check for valid email pattern and only accept .edu addresses
+  // Two letter country codes accepted before .edu (e.g. string@string.uk.edu)
+  let emailRegex = new RegExp(
+    "(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\\.)+(?:[A-Z]{2}|edu)\\b"
+  );
+
+  return emailRegex.test(email);
+}
+
+function isValidUniversityId(id) {
+  checkParamPresent(id);
+  checkIsString(id);
+  id = cleanUpString(id);
+  checkStringLength(id);
+  validateObjectId(id);
+
+  return id;
 }
 
 module.exports = {
+    stringifyId,
+    validateObjectId,
     checkArgumentLength,
+    checkParamPresent,
     checkIsString,
-    lower
+    cleanUpString,
+    checkStringLength,
+    isValidEmail,
+    isValidUniversityId
 };
